+++ src/ft_strsub.c
@@ -13,7 +13,7 @@
 #include <stdlib.h>
 #include "libft.h"
 
-char	*ft_strsub(char const *s, unsigned int start, size_t len)
+char	*ft_strsub(char const *s, size_t start, size_t len)
 {
 	char			*substr;
 	size_t			s_len;
